when the 6 hour time slice has passed for late data, recompute the 6 hour data
This commit is best explained with an example. Suppose the current time is
06:12. We receive a measurement report with data having a timestamp of 05:50.
Aggregation will run again at 07:00. The 6 hour time slice for the late data is
00:00 to 06:00, and that time slice has already passed. When aggregation runs,
we not only compute the 1 hour aggregate for the data inserted at 05:50, we
also recompute the 6 hour aggregate so that the late data is included.