making schema changes to metrics_cache_index table
Some columns have been renamed and a new column has been added. We now have the
following timestamp columns - day, collection_time_slice, and insert_time_slice.
collection_time_slice is the timestamp provided by the reported data.
insert_time_slice is the raw time slice in which the data is inserted. For
example, if data is inserted at 17:02, insert_time_slice will be 17:00. And
day is the 24 hour time slice of the collection timestamp.

insert_time_slice is the new column. Previously there was a scenario in which
we could fail to aggregate data. Suppose aggregation for a cache partition
failed. Its cache index entry would be left intact., and note that the set of
schedule ids was empty. Then late data is reported in the next time slice. At
that point, the cache index entry is updated with the schedule ids of the late
data. Prior to this commit we would have only aggregated data for the schedule
ids with late data.

Now we always update the schedule ids for the cache index, even for current
data. When we have past/late data, we now compare collection_time_slice with
insert_time_slice to determine if we can pull data from metrics_cache.